extract core features from bin
